requirements.txt
instagrapi
google-genai
python-dotenv
requests
json-repair